+++ src/components/explore/MemesByTag/MemesByTag.tsx
@@ -1,5 +1,5 @@
 import { useGetMemesByTag } from "@/application/hooks";
-import { InfiniteMemeList } from "@/components/meme";
+import { InfiniteMemeList } from "@/features/common";
 
 import { EmptyMemesView } from "../EmptyMemesView";
