--- src/components/meme/ActionSheet/index.ts
@@ -1,1 +1,0 @@
-export * from "./MemeActionSheet";
