--- src/components/meme/InfiniteMemeList/index.ts
@@ -1,1 +1,0 @@
-export * from "./InfiniteMemeList";
