--- src/components/meme/MemeInfo/Button/index.ts
@@ -1,4 +1,0 @@
-export * from "./ClipboardCopyButton";
-export * from "./CollectionSaveButton";
-export * from "./KakaoShareButton";
-export * from "./NativeShareButton";
