--- src/components/meme/MemeInfo/MemeCTAList.tsx
@@ -1,35 +1,0 @@
-import { Suspense } from "react";
-
-import { Button } from "@/common/components/Button";
-import { Icon } from "@/common/components/Icon";
-import { useOverlay } from "@/common/hooks";
-import { MemeShareModal } from "@/components/meme/MemeInfo/Modal";
-
-import { CollectionSaveButton } from "./Button";
-
-interface Props {
-  id: string;
-}
-
-export const MemeCTAList = ({ id }: Props) => {
-  const overlay = useOverlay();
-
-  return (
-    <div className="flex w-full gap-10 py-40">
-      <Button
-        className="ga-meme-share-click h-52 w-52 shrink-0 rounded-10 bg-gray-900 active:bg-black"
-        onClick={() => {
-          overlay.open(({ isOpen, close }) => (
-            <Suspense>
-              <MemeShareModal id={id} isOpen={isOpen} onClose={close} />
-            </Suspense>
-          ));
-        }}
-      >
-        <Icon height={24} name="memeShare" stroke="white" width={24} />
-      </Button>
-
-      <CollectionSaveButton id={id} />
-    </div>
-  );
-};
