--- src/components/meme/MemeItem/index.ts
@@ -1,1 +1,0 @@
-export * from "./MemeItem";
