+++ src/features/common/components/index.ts
@@ -1,4 +1,3 @@
-export * from "./ActionSheet";
 export * from "./InfiniteMemeList";
-export * from "./MemeInfo";
+export * from "./MemeActionSheet";
 export * from "./MemeItem";
