+++ src/features/memes/components/MemeDetail/index.ts
@@ -1,0 +1,1 @@
+export * from "./MemeDetail";
