+++ src/features/memes/components/index.ts
@@ -1,5 +1,6 @@
 export * from "./MemeCTAList";
 export * from "./MemeDetail";
+export * from "./MemeShareModal";
 export * from "./MemeTagList";
 export * from "./RelativeMemeList";
 export * from "./Skeleton";
